Playing with While Loops
I think I got down the study drills to do with while loops, but the last study drill asks us to write the same function with a for loop and range. I think I do this correctly, but it seemed like we did not need to use the incrememtor in the middle anymore. I was abit confused about this part. Please check and let me know what you think.